Move `FakeAccountManager` to `:app:testing`
