+++ app/common/build.gradle.kts
@@ -35,6 +35,8 @@ dependencies {
     // Required for DependencyInjectionTest to be able to resolve OpenPgpApiManager
     testImplementation(projects.plugins.openpgpApiLib.openpgpApi)
 
+    testImplementation(projects.app.testing)
+
     testImplementation(libs.robolectric)
 }
