+++ app/common/src/test/java/com/fsck/k9/account/AccountStateLoaderTest.kt
@@ -10,6 +10,7 @@ import com.fsck.k9.Identity
 import com.fsck.k9.mail.AuthType
 import com.fsck.k9.mail.ConnectionSecurity
 import com.fsck.k9.mail.ServerSettings
+import com.fsck.k9.preferences.FakeAccountManager
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
